scripts/requirements.txt
playwright==1.40.0
python-dotenv==1.0.0
supabase==2.0.3
python-jose==3.3.0
loguru==0.7.2
